pyproject.toml
[project]
name = "mikucast-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MikuCast", email = "mikucast.official@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dynaconf>=3.2.11",
    "httpx>=0.28.1",
    "jmespath>=1.0.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-ai-slim[openai]>=0.3.4",
    "questionary>=2.1.0",
    "rich>=14.0.0",
    "toml>=0.10.2",
    "typer>=0.16.0",
]

[project.scripts]
mikucast = "cli.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/cli"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "types-toml>=0.10.8.20240310",
]

[[tool.mypy.overrides]]
module = "dynaconf.*"
ignore_missing_imports = true
